Rename ZConnectionState to ClientConnectionState
